requirements.txt
Django==1.9
django-filter==1.0.4
django-tastypie==0.14.0
djangorestframework==3.6.4
Markdown==2.6.9
python-dateutil==2.6.1
python-mimeparse==1.6.0
six==1.10.0
